Move services to utils folder
