--- mainapp/services.py
@@ -1,48 +1,0 @@
-from django.db.models import QuerySet
-
-
-def all_objects(model, select_related: str = None, prefetch_related: str = None) -> QuerySet:
-    """Get all objects for a given model.
-
-     Args:
-     - model: The model for which to retrieve objects
-     - select_related: Related fields to select in the query
-     - prefetch_related: Related fields to prefetch in the query
-
-     """
-    all_objects = model.objects.all()
-    if select_related:
-        all_objects = all_objects.select_related(select_related)
-    if prefetch_related:
-        all_objects = all_objects.prefetch_related(prefetch_related)
-    return all_objects
-
-
-def get_object(model, **kwargs):
-    """
-      Get a single object of the given model based on provided criteria
-
-      Args:
-      - model: The model for which to get filtered objects
-      - **kwargs: Filtering criteria for retrieving objects
-      """
-    return model.objects.get(**kwargs)
-
-
-def get_filter_objects(model, **kwargs) -> QuerySet:
-    """Retrieve filtered objects for the given model based on provided criteria.
-        Args:
-        - model: The model for which to retrieve filtered objects
-        - **kwargs: Filtering criteria for retrieving objects
-        """
-    return model.objects.filter(**kwargs)
-
-
-def get_objects_with_limit(model, limit: int, order_by: str = None) -> QuerySet:
-    """Retrieve a limited number of objects for the given model.
-        Args:
-        - model: The model for which to retrieve objects
-        - limit: The maximum number of objects to retrieve
-        - order_by: Field to order the objects by (optional)
-        """
-    return all_objects(model).order_by(order_by)[:limit]
